migration fix1

+++ db/migrate/20230329084610_drop_questions_time.rb
@@ -1,5 +1,5 @@
 class DropQuestionsTime < ActiveRecord::Migration[7.0]
   def change
-    drop_table :questions_times
+    drop_table :questions_times if table_exists?(:questions_times)
   end
 end
